+++ app/institutions/dashboard/-components/object-list/component-test.ts
@@ -5,6 +5,10 @@ import { setupIntl } from 'ember-intl/test-support';
 import { setupRenderingTest } from 'ember-qunit';
 import { TestContext } from 'ember-test-helpers';
 import { module, test } from 'qunit';
+import Service from '@ember/service';
+const featuresStub = Service.extend({
+    isEnabled: () => true,
+});
 
 import { OsfLinkRouterStub } from 'ember-osf-web/tests/integration/helpers/osf-link-router-stub';
 
@@ -14,13 +18,15 @@ module('Integration | institutions | dashboard | -components | object-list', hoo
     setupIntl(hooks);
 
     hooks.beforeEach(function(this: TestContext) {
+        this.owner.register('service:features', featuresStub);
         this.owner.unregister('service:router');
         this.owner.register('service:router', OsfLinkRouterStub);
         const columns = [
             {
                 name: 'Title',
-                sortKey: 'title',
+                isSortable: true,
                 getValue: () => 'Title of some object',
+                propertyPathKey: 'title',
             },
             {
                 name: 'Description',
@@ -81,6 +87,12 @@ module('Integration | institutions | dashboard | -components | object-list', hoo
 
         // The table data is not blatantly incorrect
         assert.dom('[data-test-object-table-body-row]').exists({ count: 10 }, 'There are 10 rows');
+
+        // Download buttons are present
+        await click('[data-test-download-dropdown]');
+        assert.dom('[data-test-download-csv-link]').exists('CSV download link exists');
+        assert.dom('[data-test-download-tsv-link]').exists('TSV download link exists');
+        assert.dom('[data-test-download-json-link]').exists('JSON download link exists');
     });
 
     test('the table supports filtering', async function(assert) {